pyproject.toml
[project]
name = "longdelta-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.4",
    "langchain-community>=0.3.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "numpy>=1.26.0",
    "pandas>=2.3.3",
    "plotly>=5.24.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.40.0",
    "uvicorn[standard]>=0.40.0",
]
